src/collectors/filebeat.py
"""
Filebeat 采集器
TODO: 实现 Filebeat 日志采集

开发任务:
1. 实现 Filebeat 输出监控
2. 实现 Kafka 消息发送
3. 实现采集状态管理
"""
import json
from typing import Any, Dict, Generator, Optional
from kafka import KafkaProducer
from .base import BaseCollector
from ..utils.logger import get_logger

logger = get_logger(__name__)


class FilebeatCollector(BaseCollector):
    """Filebeat 采集器"""
    
    def __init__(self, name: str = "filebeat", config: Optional[Dict[str, Any]] = None):
        """
        初始化 Filebeat 采集器
        
        Args:
            name: 采集器名称
            config: 配置字典
        """
        super().__init__(name, config or {})
        self.kafka_producer = None
        self.kafka_topic = self.config.get('kafka_topic', 'logs_raw')
        
    def start(self):
        """启动采集器"""
        # TODO: 实现启动逻辑
        try:
            # 初始化 Kafka 生产者
            self.kafka_producer = KafkaProducer(
                bootstrap_servers=self.config.get('bootstrap_servers', 'localhost:9092').split(','),
                value_serializer=lambda v: json.dumps(v, ensure_ascii=False).encode('utf-8'),
                compression_type='gzip'
            )
            self.is_running = True
            logger.info(f"Filebeat 采集器 [{self.name}] 已启动")
        except Exception as e:
            logger.error(f"启动 Filebeat 采集器失败：{e}")
            raise
    
    def stop(self):
        """停止采集器"""
        # TODO: 实现停止逻辑
        self.is_running = False
        if self.kafka_producer:
            self.kafka_producer.close()
        logger.info(f"Filebeat 采集器 [{self.name}] 已停止")
    
    def collect(self) -> Generator[Dict[str, Any], None, None]:
        """
        从 Filebeat 采集日志
        
        Yields:
            日志记录
        """
        # TODO: 实现采集逻辑
        logger.info("Filebeat 采集器等待日志数据...")
        while self.is_running:
            yield {
                'status': 'running',
                'collector': self.name,
                'message': 'Waiting for logs from Filebeat'
            }
    
    def send_to_kafka(self, log_data: Dict[str, Any]):
        """
        发送日志到 Kafka
        
        Args:
            log_data: 日志数据
        """
        # TODO: 实现 Kafka 发送逻辑
        if self.kafka_producer:
            try:
                future = self.kafka_producer.send(self.kafka_topic, value=log_data)
                future.get(timeout=10)
                logger.debug(f"日志已发送到 Kafka: {log_data.get('log_type')}")
            except Exception as e:
                logger.error(f"发送日志到 Kafka 失败：{e}")
